Some shared methods ExternalInvoking and VulnerableDependency moved in CEUtils
